LU-11830 lov: avoid signed vs. unsigned comparison
In the expansion of do_div64() GCC complains about pointer comparison
because loff_t is not a u64 variable as it should be.  lov_do_div64()
also has signed vs. unsigned comparisons due to a signed loff_t.
Change lov_do_div() to use a 64-bit variable for do_div() instead of
loff_t to avoid these warnings.

Change OST_MAXREQSIZE and friends to be consistently unsigned values
to avoid compiler warnings.

Fix "lfs mirror resync" to avoid comparing signed and unsigned valued.

Lustre-change: https://review.whamcloud.com/33921
Lustre-commit: 632b3591b6ea21e3029299ef875179a4b5ca918f

Reviewed-on: https://review.whamcloud.com/34558
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/tests/mirror_io.c
@@ -364,7 +364,7 @@ ssize_t mirror_resync_one(int fd, struct llapi_layout *layout,
 {
 	uint64_t mirror_end = 0;
 	ssize_t result = 0;
-	size_t count;
+	uint64_t count;
 
 	if (end == OBD_OBJECT_EOF)
 		count = OBD_OBJECT_EOF;
@@ -373,7 +373,7 @@ ssize_t mirror_resync_one(int fd, struct llapi_layout *layout,
 
 	while (count > 0) {
 		uint32_t src;
-		size_t to_copy;
+		uint64_t to_copy;
 		ssize_t copied;
 
 		src = llapi_mirror_find(layout, start, end, &mirror_end);
